Переместил createElement, renderElement, unRenderElement в AbstractComponent
