+++ src/components/utils.js
@@ -49,58 +49,3 @@ export const getMonthFromTimeStamp = (timestamp) => MONTH.get(Number(dateObgFrom
  * @return {string}
  */
 export const getTimeFromTimeStamp = (timestamp) => dateObgFromTimestamp(timestamp).toLocaleString(locales, timeFormat);
-
-/**
- * Функция для изготовления DOM-элемента из строки
- *
- * @param {string} template HTML-код
- *
- * @return {ChildNode}
- */
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-/**
- * Функция для добавления DOM-элементов на страницу
- *
- * @param {string|Element} container CSS-селектор контейнера, в который необходимо добавить DOM-элемент
- * @param {Element} element DOM-элемент, который нужно добавить в страницу
- * @param {"append"|"prepend"|"insertBefore"|"insertAfter"} position Позиция вставки элемента, относительно контейнера, в который он вставляется
- */
-export const renderElement = (container, element, position = `append`) => {
-  let parentContainer;
-
-  if (typeof container === `string`) {
-    parentContainer = document.querySelector(container);
-  } else {
-    parentContainer = container;
-  }
-
-  switch (position) {
-    case `append`:
-      parentContainer.append(element);
-      break;
-    case `prepend`:
-      parentContainer.prepend(element);
-      break;
-    case `insertBefore`:
-      parentContainer.parentNode.insertBefore(element, parentContainer);
-      break;
-    case `insertAfter`:
-      parentContainer.parentNode.insertBefore(element, parentContainer.nextSibling);
-      break;
-  }
-};
-
-/**
- * Функция для удаления DOM-элемента со страницы
- *
- * @param {Element} element DOM-элемент, который нужно удалить со страницы
- */
-export const unRenderElement = (element) => {
-  element.parentNode.removeChild(element);
-};
